producer/requirements.txt
google-cloud-pubsub==2.21.4
google-auth==2.29.0
